refactor: tidy code
