+++ crates/cli/src/cli/shellinit.rs
@@ -1,6 +1,7 @@
-use crate::cli::shellinit_scripts;
 use anyhow::Result;
 use clap::{ArgMatches, Command};
+
+use crate::cli::shellinit_scripts::{BASH, FISH, ZSH};
 
 pub fn cli() -> Command {
     Command::new("shellinit")
@@ -10,23 +11,23 @@ pub fn cli() -> Command {
 pub fn exec(args: &ArgMatches) -> Result<()> {
     _ = args;
 
-    let is_fish = std::env::var_os("FISH_VERSION").is_some();
     let shell = std::env::var("SHELL").unwrap_or_default();
+    let is_fish = std::env::var_os("FISH_VERSION").is_some();
     let is_zsh = std::env::var_os("ZSH_VERSION").is_some() || shell.ends_with("/zsh");
 
     if is_fish {
-        let fish_script = shellinit_scripts::FISH;
+        let fish_script = FISH;
         println!("{}", fish_script);
         return Ok(());
     }
 
     if is_zsh {
-        let zsh_script = shellinit_scripts::ZSH;
+        let zsh_script = ZSH;
         println!("{}", zsh_script);
         return Ok(());
     }
 
-    let bash_script = shellinit_scripts::BASH;
+    let bash_script = BASH;
     println!("{}", bash_script);
 
     Ok(())
